// src/app/models/system/comlocation.ts
export class ComlocationModel {
    constructor() {
    }
    company_code: string = "";
    comlocation_id: string = "1";

    comlocation_code: string = "";
    comlocation_name_th: string = "";
    comlocation_name_en: string = "";

    modified_by: string = "";
    modified_date: string = "";
    flag: boolean = false;

    index: number = 0;
    select: boolean = false;

  }